callbacks extracted, CableReady processing extracted
